wip: redo server
