--- src/Location.hpp
@@ -1,29 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   Location.hpp                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: clballes <clballes@student.42barcel>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/04 11:56:41 by clballes          #+#    #+#             */
-/*   Updated: 2023/12/04 11:56:44 by clballes         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef __LOCATION_HPP__
-#define __LOCATION_HPP__
-
-#include <iostream>
-
-struct Location
-{
-    std::string path;
-    std::string alias;
-    std::string root;
-    std::string index;
-    std::string allow_methods;
-    std::string cgi_pass;
-    int         client_size;
-};
-
-#endif
